Rename files and finish up
