+++ moosebot/bot.py
