+++ moosebot/cogs/colour.py
