+++ moosebot/cogs/dad.py
