+++ moosebot/cogs/fun.py
