+++ moosebot/cogs/guess_game.py
