+++ moosebot/cogs/images.py
