+++ moosebot/cogs/info.py
