+++ moosebot/cogs/misc.py
