+++ moosebot/cogs/moderation.py
