+++ moosebot/cogs/pet.py
