+++ moosebot/converters/__init__.py
@@ -1,10 +1,9 @@
-from moosebot.converters.VoiceChannel import VoiceChannel
-from moosebot.converters.FullMember import FullMember
-from moosebot.converters.MemberDisplayName import MemberDisplayName
-from moosebot.converters.PartialMember import PartialMember
-from moosebot.converters.Role import Role
-from moosebot.converters.Url import Url
-
+from moosebot.converters.full_member import FullMember
+from moosebot.converters.member_display_name import MemberDisplayName
+from moosebot.converters.partial_member import PartialMember
+from moosebot.converters.role import Role
+from moosebot.converters.url import Url
+from moosebot.converters.voice_channel import VoiceChannel
 
 __all__ = [
     "FullMember",
