+++ moosebot/converters/partial_member.py
