+++ moosebot/converters/role.py
